Created Validation class.
